+++ packages/vira/src/Vira/Page/ToolsPage/View.hs
@@ -1,0 +1,121 @@
+{-# LANGUAGE OverloadedRecordDot #-}
+
+-- | Tool card rendering
+module Vira.Page.ToolsPage.View (
+  viewTool,
+) where
+
+import Attic.Config (AtticConfig (..), AtticServerConfig (..))
+import Data.Map.Strict qualified as Map
+import Data.Text qualified as T
+import GH.Auth.Status (AuthStatus (..))
+import Lucid
+import Vira.Page.ToolsPage.Tool
+import Vira.Widgets.Alert qualified as W
+import Vira.Widgets.Card qualified as W
+import Web.TablerIcons.Outline qualified as Icon
+
+-- | Tool display styling
+data ToolDisplay = ToolDisplay
+  { initial :: Text
+  , bgClass :: Text
+  , textClass :: Text
+  }
+
+-- | Get display styling for a tool
+viewToolDisplay :: Tool info -> ToolDisplay
+viewToolDisplay = \case
+  Attic -> ToolDisplay {initial = "A", bgClass = "bg-indigo-100", textClass = "text-indigo-600"}
+  GitHub -> ToolDisplay {initial = "G", bgClass = "bg-green-100", textClass = "text-green-600"}
+  Omnix -> ToolDisplay {initial = "O", bgClass = "bg-purple-100", textClass = "text-purple-600"}
+  Git -> ToolDisplay {initial = "G", bgClass = "bg-orange-100", textClass = "text-orange-600"}
+  Cachix -> ToolDisplay {initial = "C", bgClass = "bg-blue-100", textClass = "text-blue-600"}
+
+-- | View a tool card with its metadata and runtime info
+viewTool :: (Monad m) => Tool info -> info -> HtmlT m ()
+viewTool tool info = do
+  let meta = toolMeta tool
+      disp = viewToolDisplay tool
+  W.viraCard_ [class_ "p-6"] $ do
+    div_ [class_ "flex items-start mb-4"] $ do
+      span_ [class_ $ "h-12 w-12 mr-4 " <> disp.bgClass <> " rounded-lg flex items-center justify-center " <> disp.textClass <> " font-bold text-xl"] $
+        toHtml disp.initial
+      div_ [class_ "flex-1"] $ do
+        h3_ [class_ "text-xl font-bold text-gray-900 mb-2"] $ toHtml meta.name
+        p_ [class_ "text-gray-600 text-sm mb-3"] $ toHtml meta.description
+        div_ [class_ "mb-3 space-y-1"] $ do
+          forM_ meta.binPaths $ \binPath ->
+            code_ [class_ "block text-xs bg-gray-100 text-gray-700 px-2 py-1 rounded font-mono"] $ toHtml binPath
+
+        -- Render tool-specific info
+        viewToolInfo tool info
+
+        a_
+          [ href_ meta.url
+          , target_ "_blank"
+          , class_ "inline-flex items-center gap-1 text-indigo-600 hover:text-indigo-800 text-sm font-medium"
+          ]
+          $ do
+            span_ "Learn more"
+            span_ [class_ "w-4 h-4 flex items-center"] $ toHtmlRaw Icon.external_link
+
+-- | View tool-specific runtime information
+viewToolInfo :: (Monad m) => Tool info -> info -> HtmlT m ()
+viewToolInfo Attic cfg = do
+  div_ [class_ "mb-3"] $ do
+    case cfg of
+      Left err -> do
+        W.viraAlert_ W.AlertError $ do
+          p_ [class_ "text-red-800 font-semibold mb-1"] "✗ Parse error"
+          p_ [class_ "text-red-700 text-sm"] $ toHtml (show err :: String)
+      Right Nothing -> do
+        W.viraAlert_ W.AlertWarning $ do
+          p_ [class_ "text-yellow-800 mb-1"] "⚠ Not configured"
+          p_ [class_ "text-yellow-700 text-sm"] $ do
+            "Config file not found at "
+            code_ [class_ "bg-yellow-100 px-1 rounded"] "~/.config/attic/config.toml"
+      Right (Just atticCfg) -> do
+        W.viraAlert_ W.AlertSuccess $ do
+          case atticCfg.defaultServer of
+            Just defServer -> do
+              p_ [class_ "text-green-800 font-semibold mb-1"] $ do
+                "✓ Default server: "
+                strong_ $ toHtml defServer
+            Nothing -> do
+              p_ [class_ "text-green-800 font-semibold mb-1"] "✓ Configured"
+
+          -- Display all configured servers
+          unless (Map.null atticCfg.servers) $ do
+            p_ [class_ "text-green-700 text-xs mt-2 mb-1"] "Configured servers:"
+            div_ [class_ "space-y-1"] $ do
+              forM_ (Map.toList atticCfg.servers) $ \(serverName, serverCfg) -> do
+                div_ [class_ "text-green-700 text-xs pl-2"] $ do
+                  strong_ $ toHtml serverName
+                  ": "
+                  code_ [class_ "bg-green-100 px-1 rounded"] $ toHtml serverCfg.endpoint
+                  case serverCfg.token of
+                    Just _ -> span_ [class_ "ml-2 text-green-600"] "🔑"
+                    Nothing -> span_ [class_ "ml-2 text-yellow-600"] "⚠ No token"
+viewToolInfo GitHub status = do
+  div_ [class_ "mb-3"] $ do
+    case status of
+      Authenticated {host, login, scopes} -> do
+        W.viraAlert_ W.AlertSuccess $ do
+          p_ [class_ "text-green-800 font-semibold mb-1"] $ do
+            "✓ Authenticated as "
+            strong_ $ toHtml login
+            " on "
+            strong_ $ toHtml host
+          p_ [class_ "text-green-700 text-xs"] $ do
+            "Scopes: "
+            toHtml $ T.intercalate ", " scopes
+      NotAuthenticated -> do
+        W.viraAlert_ W.AlertError $ do
+          p_ [class_ "text-red-800 mb-1"] "✗ Not authenticated"
+          p_ [class_ "text-red-700 text-sm"] $ do
+            "Run "
+            code_ [class_ "bg-red-100 px-1 rounded"] "gh auth login"
+            " to authenticate."
+viewToolInfo Omnix () = mempty
+viewToolInfo Git () = mempty
+viewToolInfo Cachix () = mempty
