+++ packages/vira/vira.cabal
@@ -59,8 +59,9 @@ library
       Vira.Page.RegistryPage
       Vira.Page.RepoPage
       Vira.Page.SettingsPage
-      Vira.Page.ToolsPage
+      Vira.Page.ToolsPage.Handler
       Vira.Page.ToolsPage.Tool
+      Vira.Page.ToolsPage.View
       Vira.State.Acid
       Vira.State.Core
       Vira.State.JSON
